requirements.txt
streamlit
transformers
torch
pillow
pytesseract
pdf2image
PyPDF2